mcp clientの責務を接続・ツール・リトライ・表示に分割
